Workaround for a bug in gmake 3.81.
Previously, this would work fine on my system, because I have make 3.82.
However, the CS systems run 3.81 and seem to struggle with having an
equals sign in a `define`. The documentation [1] says that this should
work.

Consequently, we didn't build the list of targets to make properly on
the affected systems.

The fix is also valid Make code, so it should work fine in 3.81 and
3.82.

[1] http://www.gnu.org/software/make/manual/make.html#Multi_002dLine