FancyCat: Fix compiler warnings.
This fixes warnings about 0xFF causing overflows on signed chars.

The warning was masked before because FancyCat wasn't being built with
the correct CFLAGS. As it turns out, the Makefile in build/ hasn't been
executed at all since the reorganization!

This also includes a change to do a recursive submake to build/ rather
than building BuildImage and FancyCat directly from the toplevel
makefile.