ext2: No need to store the mountpoint.
The VFS layer should translate paths for the ext2 driver. Every ext2
mount should be given paths (for raw_read(), etc.) which are realative
to the root of that filesystem, not the operating system.

So, if we have a filesystem mounted at /food and we want to read
/food/egg_bagel_recipe.txt, the ext2 driver at that mountpoint will only
see a request for /egg_bagel_recipe.txt.