Many changes:
The ELF loader has a few more definitions and the exec() commands have been
restructured to take a program path instead. Currently the kernel doesn't
support processes until the ELF loader is finished. The kernel currently
halts inside of the _init() function.

The networking code is coming along. Not much has been added other than some
documentation and stubs for functions.

Added another method to the PCI module, __pci_find_device() which takes a
vendor and device id. Also, some changes were made to the type_0 struct. First
the BAR addresses are now an array of 6 32-bit integers. Also there is a new
field that contains the size in bytes of each of the BARs memory locations.
These are also stored in an array of 6 32-bit integers.

Inside of the user_programs directory a new file crt0.S has been added, this
will be the startup code for C programs...eventually

Ulib was modified to use the new definition of exec()