ext2: Fix file lookup in directories.
There was a bug which allowed directory entries to match if the
directory entry we were currenly iterating over was a substring of the
file we wanted to find.

This patch fixes that bug by making sure that the lengths of the two
filenames are the same before calling strncmp() on them, which should
eleminate matching on a substring.

Note that we can't just use strcmp() here or rely on strncmp()
terminating early if it reaches the end of one file -- directory entries
don't have null terminated filenames.