ext2: Correct directory traversal for filenames.
Fixes Github issue #4: Directory traversals don't fail properly.

The underlying cause was that the documentation incorrectly described the way
that we know when we've run out of entries to read in a directory. The
documentation says that we know we've reached the end when the directory entry
we have refers to inode 0. In libext2fs, the last directory entry has its size
set so that it takes up the remainder of the inode.

Now we test both conditions to see if we've finished traversing the directory.

A new set of tests 'test_file_lookup' and 'test_nonexistant_file_lookup'
confirms that this fix works.